requirements.txt
asgiref==3.5.2
Django==4.1.1
django-bootstrap-form==3.4
django-filter==22.1
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.6
